+++ pkg/store/tsdb_test.go
@@ -34,7 +34,20 @@ func TestTSDBStore_Info(t *testing.T) {
 
 	testutil.Equals(t, []storepb.Label{{Name: "region", Value: "eu-west"}}, resp.Labels)
 	testutil.Equals(t, storepb.StoreType_RULE, resp.StoreType)
-	testutil.Equals(t, int64(0), resp.MinTime)
+	testutil.Equals(t, int64(math.MaxInt64), resp.MinTime)
+	testutil.Equals(t, int64(math.MaxInt64), resp.MaxTime)
+
+	app := db.Appender()
+	_, err = app.Add(labels.FromStrings("a", "a"), 12, 0.1)
+	testutil.Ok(t, err)
+	testutil.Ok(t, app.Commit())
+
+	resp, err = tsdbStore.Info(ctx, &storepb.InfoRequest{})
+	testutil.Ok(t, err)
+
+	testutil.Equals(t, []storepb.Label{{Name: "region", Value: "eu-west"}}, resp.Labels)
+	testutil.Equals(t, storepb.StoreType_RULE, resp.StoreType)
+	testutil.Equals(t, int64(12), resp.MinTime)
 	testutil.Equals(t, int64(math.MaxInt64), resp.MaxTime)
 }
 
@@ -277,14 +290,14 @@ func TestTSDBStore_LabelValues(t *testing.T) {
 }
 
 // Regression test for https://github.com/thanos-io/thanos/issues/1038.
-func TestTSDBStore_Series_SplitSamplesIntoChunksWithMaxSizeOfUint16_e2e(t *testing.T) {
+func TestTSDBStore_Series_SplitSamplesIntoChunksWithMaxSizeOf120(t *testing.T) {
 	defer leaktest.CheckTimeout(t, 10*time.Second)()
 
 	db, err := e2eutil.NewTSDB()
 	defer func() { testutil.Ok(t, db.Close()) }()
 	testutil.Ok(t, err)
 
-	testSeries_SplitSamplesIntoChunksWithMaxSizeOfUint16_e2e(t, db.Appender(), func() storepb.StoreServer {
+	testSeries_SplitSamplesIntoChunksWithMaxSizeOf120(t, db.Appender(), func() storepb.StoreServer {
 		tsdbStore := NewTSDBStore(nil, nil, db, component.Rule, labels.FromStrings("region", "eu-west"))
 
 		return tsdbStore